load preferences options from configuration file
Also adjusts to support the null noteLimit (i.e. no limit).
The right way to go about this would probably be not to set the upper value to 127 but to base
this value on the number of key of the keyboard and possible transpositions
(both have to become supported MIDI options first).