Challenge: add limit of time and number of heats
The "speeds" values are improvised based on observation in jalmus, because the actual
values are not immediately obvious in the source - seems to be more built around the
animation (which we don't have [yet] than about fixed times; cf.
https://sourceforge.net/p/jalmus/code/ci/master/tree/src/net/jalmus/Jalmus.java#5654

Also adds a stop and reset functionality which one needs to be in control of the
lifecycle of the game.