AnimatedAbcNotation: introduce jalmus look-alike game
Introduces a component to modify the default abcjs output to emulate the
visual impression known from the "note reading" "normal game" - with the
note to play slowy making its way from left to right, illustrating the
remaining time until the next challenge.

Maybe it is because of vue 3 but listening to child component life cycle
events was not possible the way some examples indicated (cf. [1]) so a
dedicated event (which feels redundant) was introduced.

[1]: https://stackoverflow.com/questions/50635199/vue-js-2-0-child-component-mounted-callback