Challenge: track and show current heat
Shows the current heat and uses it to decide if we reached the last one.

Also leverages the heat number to build a key for the rendered abc,
resulting in its redrawn for each heat. At the time of writing
this feels like the cleanest way of e.g. starting animations inside of
abc over for every note (as opposed to fighting with reactivity by e.g.
setting its abc prop to empty for a brief moment or comparable). cf.
[1]. In practice this will be less of a concern once games won't show
the same challenge twice in row but it's good to know the system is
robust enought to be able to handle it no less.

[1]: https://michaelnthiessen.com/force-re-render/