configure preferences via schema, formvuelate & primevue
Uses https://www.npmjs.com/package/formvuelate to build the preferences
form based on https://www.npmjs.com/package/primevue components.

Unfortunately I could not find an out-of-the-box form builder for vue 3
which already is opinionated to use a certain component lib or provides
its own basic components. The poor src/components/form/* bridge
component is the result of that - hopefully it can go away once there is
some more adoption of vue 3 in the wild.

Not sure if augmenting the schema the way it happens here is wise but it
certainly takes away a lot of redundancy (specifying identifiers for
labels, placeholders time and again) and while the formvuelate lookup
plugin[0] looks like it is suited to come into play there, I could not
find an easy way to access the current prop name (model, in case of the
object way of writing the schema) or component members (e.g. $i18n).

[0]: https://formvuelate.js.org/#lookup-plugin