SchemaBasedForm: use locally registered components
Using the newly documented[0] way of registering form components locally
instead of globally.

Not sure if and what the documented use of `markRaw`[1] does for us but
it does not seem to hurt either. FWIW I did not see a warning like "Vue
received a Component which was made a reactive object.", as mentioned on
the web in connection with it, without it.

This is how I imagined 44c9f87 to turn out. Filing [2] now made local
component registration possible.

[0]: https://formvuelate.js.org/#using-locally-imported-components
[1]: https://v3.vuejs.org/api/basic-reactivity.html#markraw
[2]: https://github.com/formvuelate/formvuelate/issues/144