MidiCapability: make it explicit we don't need SysEx
"Web MIDI use is classified into two groups, one is non-privilege use,
and the other is privilege use with sysex permission."[0]

We only need access to the former group - let's make this explicit.

This caught my eye as Chromium 87 kept showing a deprecation warning
"Web MIDI will ask a permission to use even if the sysex is not
specified in the MIDIOptions since around M82, around May 2020". The
user agent prompts for user permission to use MIDI devices in case SysEx
is requested (which makes sense given the security implications[1]),
otherwise does not and shows this warning instead at the moment. This
seems to be the migration path the chrome developers picked and nothing
can be done about the message at the moment, even though I think we are
not doing anything wrong on our end and would be happy for the
permission prompt even for non-SysEx requests.

[0]: https://www.chromestatus.com/feature/5138066234671104
[1]: https://webaudio.github.io/web-midi-api/ see "Security and Privacy Considerations of MIDI"