add virtual keyboard
The setup (octaves) shown on the virtual keyboard happen to match my
physical keyboard but should be made configurable.

In a next step the physical (midi) keyboard should probably be made
optional so that jaslmus can also be used without a physical input
device (even though this is somewhat besides the point).

Unfortunately, when mounting VirtualKeyboard in jest, there is an error
when triggering piano-chart's create() which pretty much prevents
testing this component: "Getting bbox of element "text" is not possible.
TypeError: e.node.getBBox" - maybe that's a limitation of jest dom when
it comes to working with SVG?