make midi capability optional
Allow for users with incompatible browsers / restrictive permissions to
exclusively use the virtual keyboard.

This increases App's complexity quite a bit, as it now contains code
paths that rely on midi support, and others which omit it. I did not
find a more elegant way of solving that which did not involve forking at
a higher level and then duplicating the similarities. Maybe we could
convert the user experience a to step-like approach, first checking midi
support, building the list of keyboards from there and only then
starting the challenges, … with them injected.