distinguish between types of keyboard events (on, off)
I wanted the addListener/removeListener methods to map the type to the
allowed KeyboardListener flavor (i.e. taking listners for only on or off
events respectively) but somehow it did not work out - methods were fine
but the resulting listener maps would refuse to receive
KeyboardListener<T> even though I think the type show help narrow which
scenario is happening – not sure how to make this happen. It's only a
theoretical issue as long as all KeyboardEventTypes implement the same
base event.