play sound when keys are struck
Hook up a (tone.js) sampler to our DispatchingKeyboard (gathering both
virual and physical input) and feed it with some samples (known from
MIDI.js).

The (tone.js) sampler implements release of a key without respect for
dimension – this does not seem to be academically correct but bearable.
Our events are set up in a way that we _do_ handle velocity, we just
don't pass it on in this case.

I would have liked to get the sound font from a package but the repo is
huge and not available as an npm package, so implement some download
"magic" instead.

This also touches linting as the newly added build folder was not picked
up with eslint's default configuration.