OptionalMidiCapability: remove midiError from emits list
This slipped in during 98f48dc but no such event was actually ever
emitted. I decided against implementing it to eliminate the resulting
complexity as App's behavior in this case is to swallow the error. In
theory there could be use for knowing about the error, but that's not
the case at the moment.